refactor: merge SDP pipeline generator into single sdk_generate_template_job.py
Removes sdk_generate_sdp_pipeline.py; pipeline generation is now part of
the same script and output file (resources/jobs.yml). One generator,
one output, one call in Makefile and CI.
